平均值计算可能溢出

+++ servant/servant/PropertyReport.h
@@ -82,7 +82,7 @@ public:
     protected:
         void clear()                { _sum = 0; _count = 0; }
     private:
-        int _sum;
+        long _sum;
         int _count;
     };
